align indices and token counters; minor refactor
