+++ models/demos/deepseek_v3_d_p/tests/pcc/test_prefill_combine.py
@@ -30,8 +30,8 @@ from models.demos.deepseek_v3_d_p.tt.moe.init_helpers import (
     initialize_predictable_test_inputs,
     initialize_test_inputs,
 )
-from models.demos.deepseek_v3_d_p.tt.moe.moe_gate_prefill2d import MoERoutingSetup as TtMoeRoutingSetup
 from models.demos.deepseek_v3_d_p.tt.moe.tt_combine import TtCombineModule
+from models.demos.deepseek_v3_d_p.tt.moe.tt_moe_routing_setup import TtMoERoutingSetup
 from models.demos.deepseek_v3_d_p.tt.moe.validation_helpers import (
     assert_output_shape,
     log_combine_mismatch_details,
@@ -260,16 +260,6 @@ def test_ttnn_combine(
         )
         logger.debug("Using RANDOM test data")
 
-    # Compute gate outputs before dispatch (same for all EP ranks since indices are shared)
-    expert_offsets, expert_token_counts, _ = get_gate_outputs(
-        indices,
-        dispatch_group_size,
-        num_routed_experts,
-        experts_per_chip,
-        seq_len_per_chip,
-        num_experts_per_tok,
-    )
-
     # Create expert dispatch table
     expert_dispatch_table = ExpertMapping.create_dispatch_table(
         num_routed_experts=num_routed_experts,
@@ -283,8 +273,19 @@ def test_ttnn_combine(
         num_routed_experts=num_routed_experts,
     )
 
+    # Compute gate outputs before dispatch (same for all EP ranks since indices are shared)
+    expert_offsets, expert_token_counts, _ = get_gate_outputs(
+        indices,
+        dispatch_group_size,
+        num_routed_experts,
+        experts_per_chip,
+        seq_len_per_chip,
+        num_experts_per_tok,
+        expert_dispatch_table=expert_dispatch_table,
+    )
+
     # create tt-moe routing setup module
-    tt_moe_routing_setup = TtMoeRoutingSetup(
+    tt_moe_routing_setup = TtMoERoutingSetup(
         mesh_device=mesh_device, expert_dispatch_table=expert_dispatch_table, num_links=num_links
     )
 
@@ -357,9 +358,7 @@ def test_ttnn_combine(
 
     _, tt_expert_token_counts, _ = tt_moe_routing_setup(
         ttnn_top_k_experts_indices=tt_indices,
-        dispatch_group_size=dispatch_group_size,
         num_routed_experts=num_routed_experts,
-        experts_per_chip=experts_per_chip,
         seq_len_per_chip=seq_len_per_chip,
         num_experts_per_tok=num_experts_per_tok,
     )
