+++ models/demos/deepseek_v3_d_p/tests/pcc/test_prefill_dispatch.py
@@ -372,6 +372,7 @@ def test_ttnn_dispatch(
         experts_per_chip,
         seq_len_per_chip,
         num_experts_per_tok,
+        expert_dispatch_table=expert_dispatch_table,
     )
 
     # Forward pass through TTNN dispatch
