+++ models/demos/deepseek_v3_d_p/tests/torch/test_torch_dispatch_combine.py
@@ -96,6 +96,7 @@ def test_torch_dispatch_combine(
         experts_per_chip,
         seq_len_per_chip,
         num_experts_per_tok,
+        expert_dispatch_table=expert_dispatch_table,
     )
 
     # Forward pass through dispatch module
